Remove all parsing references from proof engine
