+++ proof_frog/proof_frog.py
@@ -1,6 +1,8 @@
 import sys
 from colorama import init
+import os
 from . import frog_parser
+from . import frog_ast
 from . import proof_engine
 
 
@@ -33,9 +35,32 @@ def main() -> None:
                 usage()
     elif argv[1] == "prove":
         engine = proof_engine.ProofEngine(len(argv) > 3 and argv[3] == "-v")
-        engine.prove(argv[2])
+
+        proof_file = frog_parser.parse_proof_file(argv[2])
+        for imp in proof_file.imports:
+            file_type = _get_file_type(imp.filename)
+            root: frog_ast.Root
+            match file_type:
+                case frog_ast.FileType.PRIMITIVE:
+                    root = frog_parser.parse_primitive_file(imp.filename)
+                case frog_ast.FileType.SCHEME:
+                    root = frog_parser.parse_scheme_file(imp.filename)
+                case frog_ast.FileType.GAME:
+                    root = frog_parser.parse_game_file(imp.filename)
+                case frog_ast.FileType.PROOF:
+                    raise TypeError("Cannot import proofs")
+
+            name = imp.rename if imp.rename else root.get_export_name()
+            engine.add_definition(name, root)
+
+        engine.prove(proof_file)
     else:
         usage()
+
+
+def _get_file_type(file_name: str) -> frog_ast.FileType:
+    extension: str = os.path.splitext(file_name)[1].strip(".")
+    return frog_ast.FileType(extension)
 
 
 if __name__ == "__main__":
